medclaim-ai/backend/tests/requirements.txt
# Test requirements for medclaim-ai backend
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
fastapi[all]>=0.104.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
pillow>=10.0.0
python-magic>=0.4.27
reportlab>=4.0.0
pymupdf>=1.23.0
pdfplumber>=0.9.0
langchain>=0.1.0
tiktoken>=0.5.0
requests>=2.31.0
